BE/requirements.txt
azure-functions
requests
openai>=1.0.0
azure-cosmos>=4.5.0
